13-refactor-code-to-be-more-rust-like Moved most of the code to the lib.rs crate
